Worked on navs
I created the vertical nav on the left of the page, moved the stories to the right to create a space for a vertical nav. I mostly need to work on improving both navs, primarily giving the vertical nav spacing. Also need to fix a created problem with the stories being pushed down